refactor: move coord maps into separate module
