+++ packages_rs/nextclade/src/analyze/aa_changes.rs
@@ -2,6 +2,7 @@ use crate::analyze::aa_del::AaDel;
 use crate::analyze::aa_sub::AaSub;
 use crate::analyze::nuc_del::NucDelRange;
 use crate::analyze::nuc_sub::NucSub;
+use crate::coord::coord_map2::cds_codon_pos_to_ref_range;
 use crate::gene::cds::Cds;
 use crate::gene::gene::GeneStrand;
 use crate::io::aa::Aa;
@@ -10,7 +11,6 @@ use crate::io::letter::Letter;
 use crate::io::letter::{serde_deserialize_seq, serde_serialize_seq};
 use crate::io::nuc::Nuc;
 use crate::translate::complement::reverse_complement_in_place;
-use crate::translate::coord_map2::cds_codon_pos_to_ref_range;
 use crate::translate::translate_genes::{CdsTranslation, Translation};
 use crate::utils::collections::extend_map_of_vecs;
 use crate::utils::position::{AaRefPosition, NucRefGlobalPosition, PositionLike};
