+++ packages_rs/nextclade/src/run/nextalign_run_one.rs
@@ -1,9 +1,9 @@
 use crate::align::align::align_nuc;
 use crate::align::insertions_strip::{get_aa_insertions, insertions_strip};
 use crate::align::params::AlignPairwiseParams;
+use crate::coord::coord_map::CoordMapGlobal;
 use crate::io::gene_map::GeneMap;
 use crate::io::nuc::Nuc;
-use crate::translate::coord_map::CoordMapGlobal;
 use crate::translate::translate_genes::{translate_genes, Translation};
 use crate::types::outputs::{NextalignOutputs, PeptideWarning};
 use eyre::Report;
