+++ packages_rs/nextclade/src/translate/frame_shifts_translate.rs
@@ -1,8 +1,8 @@
+use crate::coord::coord_map::CoordMapLocal;
+use crate::coord::coord_map2::cds_range_to_ref_ranges;
 use crate::gene::cds::Cds;
 use crate::io::letter::Letter;
 use crate::io::nuc::Nuc;
-use crate::translate::coord_map::CoordMapLocal;
-use crate::translate::coord_map2::cds_range_to_ref_ranges;
 use crate::utils::position::{NucAlnLocalPosition, PositionLike};
 use crate::utils::range::{AaRefRange, NucAlnLocalRange, NucRefGlobalRange, Range};
 use eyre::Report;
