+++ packages_rs/nextclade/src/types/outputs.rs
@@ -10,10 +10,10 @@ use crate::analyze::letter_ranges::{GeneAaRange, NucRange};
 use crate::analyze::nuc_del::NucDelRange;
 use crate::analyze::nuc_sub::NucSub;
 use crate::analyze::pcr_primer_changes::PcrPrimerChange;
+use crate::coord::coord_map::CoordMapGlobal;
 use crate::io::json::json_parse;
 use crate::io::nuc::Nuc;
 use crate::qc::qc_run::QcResult;
-use crate::translate::coord_map::CoordMapGlobal;
 use crate::translate::frame_shifts_translate::FrameShift;
 use crate::translate::translate_genes::Translation;
 use crate::utils::range::{AaRefRange, NucRefGlobalRange};
